Convert HistoryActivity into a Fragment
This commit renames the HistoryActivity to HistoryFragment and updates the layout file and other related classes to reflect this change.

The history.xml layout had to be migrated from a merge tag with a LinearLayout to a ConstraintLayout to correctly position the Floating Action Button (FAB).

Activities that declared HistoryActivity as their parent in the AndroidManifest have been temporarily re-parented to MainLayout. A follow-up commit should address this parenting structure more appropriately.
