+++ app/AndroidManifest.xml
@@ -117,28 +117,23 @@
         <activity
             android:name=".view.DetailActivity"
             android:configChanges="orientation|screenSize|uiMode"
-            android:parentActivityName=".view.HistoryActivity"
+            android:parentActivityName=".view.MainLayout"
             android:theme="@style/AppTheme.NoActionBar">
             <!-- Parent activity meta-data to support 4.0 and lower -->
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value=".view.HistoryActivity"/>
+                android:value=".view.MainLayout"/>
         </activity>
-
-        <activity
-            android:name=".view.HistoryActivity"
-            android:label="@string/History"
-            android:theme="@style/AppTheme.NoActionBar"/>
 
         <activity
             android:name=".view.ManualActivity"
             android:label="@string/Manual_activity_title"
-            android:parentActivityName=".view.HistoryActivity"
+            android:parentActivityName=".view.MainLayout"
             android:theme="@style/AppTheme">
             <!-- Parent activity meta-data to support 4.0 and lower -->
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value=".view.HistoryActivity"/>
+                android:value=".view.MainLayout"/>
         </activity>
 
         <activity
